desktop: refactor our code into tauri plugins
nfci
